React-Routes working correctly, began initial link navigiation, which is working
