+++ index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
 <head>
-    <title>Bubo</title>
+    <title>Quest</title>
 
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.0.0/jquery.min.js" integrity="sha384-THPy051/pYDQGanwU6poAc/hOdQxjnOEXzbT+OuUAFqNqFjL+4IGLBgCJC3ZOShY" crossorigin="anonymous"></script>
@@ -13,9 +13,8 @@
 
 </head>
 <body>
-
     <div id="app" />
-    <script src="/Public/bundle.js"></script>
+    <script src="bundle.js"> $.stellar({horizontalScroll:false}); </script>
 
 </body>
 </html>
