+++ package.json
@@ -1,10 +1,10 @@
 {
-  "name": "quest-react",
+  "name": "react-todos",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "webpack-dev-server --inline --hot"
+    "start": "webpack-dev-server"
   },
   "keywords": [],
   "author": "Eric Mok",
@@ -29,8 +29,8 @@
     "react": "^0.14.8",
     "react-bootstrap": "^0.30.3",
     "react-dom": "^0.14.8",
-    "react-hot-loader": "^1.3.0",
     "react-modal-es6": "^0.1.1",
+    "react-router": "^2.8.1",
     "sequelize": "^3.24.3",
     "sequelize-cli": "^2.4.0",
     "underscore": "^1.8.3",
