+++ src/components/missionmain.js
@@ -1,34 +1,31 @@
 import React from 'react';
+import MissionsList from './missions-list';
+import Createmission from './create-mission';
 
-import MissionsList from '../components/missions-list';
-import Createmission from '../components/create-mission';
-
-export default class MissionHome extends React.Component {
-
+export default class MissionMain extends React.Component {
     render() {
+        return (
 
-	    return (
-
-	    		<div id="missionSection">
-	            	<h1>Create a Mission</h1>
+            <div>
+                <h1>Create a Mission</h1>
+                
+                <Createmission
+                    missions={this.props.missions}
+                    createTask={this.createTask.bind(this)}
+                />
+                <MissionsList
+                    missions = {this.props.missions}
+                    toggleTask = {this.toggleTask.bind(this)}
+                    saveTask = {this.saveTask.bind(this)}
+                    deleteTask = {this.deleteTask.bind(this)}
+                />
                     
-	            	<Createmission
-                        missions={this.props.missions}
-                        createTask={this.createTask.bind(this)}
-                    />
-                    <MissionsList
-                        missions = {this.props.missions}
-                        toggleTask = {this.toggleTask.bind(this)}
-                        saveTask = {this.saveTask.bind(this)}
-                        deleteTask = {this.deleteTask.bind(this)}
-                    />
-                        
-                </div>
+            </div>
 
-	    );
-	}
+         )
+    }
 
-	// MISSION (tasks CRUD):
+    // MISSION (tasks CRUD):
     createTask(task) {
         this.props.missions.push({
             task,
@@ -53,4 +50,5 @@ export default class MissionHome extends React.Component {
         const removeTask=_.remove(this.props.missions, mission=> mission.task ===taskDelete);
         this.setState({missions: this.props.missions});
     }
+
 }
