+++ webpack.config.js
@@ -6,7 +6,7 @@ module.exports = {
     entry: [
         'webpack-dev-server/client?http://127.0.0.1:8080/',
         'webpack/hot/only-dev-server',
-        './src/Client/pages/layout'
+        './src'
     ],
     output: {
         path: path.join(__dirname, 'public'),
